adding program to move LO frequencies to minimize number of beaters.
If LO frequencies of two roaches are chosen two closely on the same feedline, aliases of the generated frequencies of one roach that are not cut out by the antialiasing filter can interfere with the readout frequencies of the other roach.  This code will move the LO frequencies randomly within an area so there are a minimum number of aliases overlapping with readout frequencies.